Development

See merge request ctd/proyecto-integrador-1022/0521-pt-c1/grupo-01!88

+++ backend/src/main/java/com/example/piBack/Service/CategoryService.java
@@ -1,7 +1,10 @@
 package com.example.piBack.Service;
 
+import com.example.piBack.Exceptions.BadRequestException;
+import com.example.piBack.Exceptions.EntityNotFoundException;
 import com.example.piBack.Model.Category;
 import com.example.piBack.Repository.CategoryRepository;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,16 +15,26 @@ import java.util.Optional;
 @Service
 public class CategoryService {
 
+    private static final Logger logger = Logger.getLogger(CategoryService.class);
     @Autowired
     CategoryRepository categoryRepository;
 
-    public Category addCategory(Category category) {
-        categoryRepository.save(category);
-        return category;
+
+
+    public Category addCategory(Category category) throws BadRequestException {
+        if(category == null || category.getDescription() == null || category.getTitle() == null) {
+            throw new BadRequestException("Los datos ingresados para registrar la categoria no son correctos o están incompletos. Debe incluir obligatoriamente: descripcion y titulo.");
+        } else {
+            categoryRepository.save(category);
+            logger.info("New category: " + category + " has been successfully saved.");
+            return category;
+
+        }
     }
 
     public Collection<Category> listCategory(){
         List<Category> categories = categoryRepository.findAll();
+        logger.info("All categories will be listed.");
         return categories;
     }
 
@@ -30,14 +43,27 @@ public class CategoryService {
         return category;
     }
 
-    public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
+    public void deleteCategory(Long id)throws EntityNotFoundException {
+        logger.info("Category with id: " + id + " will be deleted.");
+        Optional<Category> category = findCategory(id);
+        if (category.isPresent())
+            categoryRepository.deleteById(id);
+        else
+            throw new EntityNotFoundException("Category with " + id + " was not found");
+        logger.error("Category with id: " + id + "was not found in our DB.");
     }
 
-    public Optional<Category> findCategory(Long id) {
+    public Optional<Category> findCategory(Long id) throws EntityNotFoundException{
         Optional<Category> category = categoryRepository.findById(id);
+        if (category.isPresent())
+            categoryRepository.deleteById(id);
+        else
+            throw new EntityNotFoundException("Category with " + id + " was not found");
         return category;
     }
+
+
+
 
 }
